ionizer/pyproject.toml
[tool.poetry]
name = "ionizer"
version = "0.1.0"
description = ""
authors = ["ZerglingGo <zerglinggo@zerglinggo.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0"
djangorestframework = "^3.14.0"
drf-nested-routers = "^0.93.4"
drf-standardized-errors = "^0.12.6"
django-redis = "^5.4.0"
django-environ = "^0.11.2"
mysqlclient = "^2.2.1"
uvicorn = {extras = ["standard"], version = "^0.25.0"}
gunicorn = "^21.2.0"
boto3 = "^1.34.15"
django-auditlog = "^2.3.0"
django-silk = "^5.1.0"
tsidpy = "^1.1.5"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.5.0"
isort = "^5.12.0"

[tool.black]
line-length = 120

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
